chore: simplify implementation (use navigator.clipboard)
