+++ lib/index.js
@@ -1,4 +1,4 @@
-import NativeCopyPaste from './NativeCopyPaste';
+import NativeCopyPaste from './NativeCopyPaste.js';
 
 export default {
   __init__: [ 'nativeCopyPaste' ],
